% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_formula.R
\name{quadratic_formula}
\alias{quadratic_formula}
\title{Quadratic Formula}
\usage{
quadratic_formula(a, b, c)
}
\arguments{
\item{a}{coefficient of the square term in a quadratic (numeric)}

\item{b}{coefficient of linear term in a quadratic (numeric)}

\item{c}{constant term in quadretic (numeric)}
}
\value{
the roots of the quadratic, numeric if repeated roots, vector of numerics if
distinct real roots, vector of (class) complex objects if imaginary roots
}
\description{
Quadratic Formula
}
\examples{
quadratic_formula(2, -12, 18)
quadratic_formula(1, -5, 6)
quadratic_formula(2, -2, 1)
}
